feat: Add run-doc subcommand to run the test and generate markdown docs
This changeset adds a new sub-command named run-doc which can be used
to generate markdown from the steps in the .basi file as a way to aid
in documenting steps for applications in addition to testing them.

This is very early and very experimental, lot's of things to figure out.
But it works okay enough for a first pass